Add more notes for CI setup
