+++ source/commands/init/interfaces.ts
@@ -1,0 +1,29 @@
+export interface InitState {
+  filename: string
+  botName: string
+
+  isWindows: boolean
+  isMac: boolean
+  isBabel: boolean
+  isTypeScript: boolean
+  supportsHLinks: boolean
+
+  isAnOSSRepo: boolean
+
+  hasCreatedDangerfile: boolean
+  hasSetUpAccount: boolean
+  hasSetUpAccountToken: boolean
+
+  repoSlug: string | null
+  ciType: "travis" | "circle" | "unknown"
+}
+
+export interface InitUI {
+  header: (msg: String) => void
+  command: (command: string) => void
+  say: (msg: String) => void
+  pause: (secs: number) => Promise<{}>
+  waitForReturn: () => void
+  link: (name: string, href: string) => string
+  askWithAnswers: (message: string, answers: string[]) => string
+}
